@매핑 -- 신청 현황 작업 해야함

+++ src/views/Canendal/components/ShowSelectEvent.js
@@ -1,4 +1,4 @@
-import React,{useEffect,useState} from 'react';
+import React,{useEffect,useState,useRef} from 'react';
 import Button from '@material-ui/core/Button';
 import Dialog from '@material-ui/core/Dialog';
 import Chip from '@material-ui/core/Chip';
@@ -11,6 +11,8 @@ import Slider from '@material-ui/core/Slider';
 import UpdatePlan from './UpdatePlan';
 import ConfirmDialog from 'components/ConfirmDialog/ConfirmDialog';
 import * as axiosDelete from '@axios/delete';
+import * as axiosPatch from '@axios/patch';
+import MessageBox from 'components/MessageBox/MessageBox';
 
 const PrettoSlider = withStyles({
   root: {
@@ -61,12 +63,35 @@ const useStyles = makeStyles({
 
 export default function ShowSelectEvent(props) {
   const classes = useStyles();
-
-  const [deleteState,setDeleteState] = useState(false);
+  const progress = useRef();
   const [updatePlanState,setUpdatePlanState] = React.useState(false);
+  const [confirmState,setConfirmState] = useState(false);
+
   const [open, setOpen] = React.useState(props['open']);
   const [checkDate,setCheckDate] = React.useState(0);
   const {event} = props;
+  const [confirmDialogState,setConfirmDialogState] = useState(false);
+  const [confirmDialogInfo,setConfirmDialogInfo] = useState({
+    title : "",
+    content : ""
+  })
+  const [showMessageState,setShowMessageState] = useState(false);
+  const [MessageBoxState,setMessageBoxState] = useState(
+    {
+      content : "",
+      level : "success",
+      time : 2000
+    }
+  );
+
+  const _messageBoxHandle = (show,content,time,level) => {
+    setShowMessageState(show);
+    setMessageBoxState({
+      content : content,
+      time : time,
+      level : level
+    })
+  }
 
   const messageBoxHandle = (show,content,time,level) => {
     props.messageBoxHandle(show,content,time,level);
@@ -98,13 +123,50 @@ export default function ShowSelectEvent(props) {
     handleClose();
   }
 
-  const deletePlan = () => {
-    axiosDelete.deleteNotContainsData("http://localhost:8090/api/teamManage/plan/" + props['event']['groupId'],deletePlanSuccess);
+  const yseClickHandle = () => {
+    if(confirmState)
+      axiosDelete.deleteNotContainsData("http://localhost:8090/api/teamManage/plan/" + props['event']['groupId'],deletePlanSuccess);
+    else{
+      let data = new FormData();
+      data.append("progress",progress.current.value);
+      axiosPatch.patchContainsData("http://localhost:8090/api/teamManage/plan/"+ props['event']['groupId']+"/updateProgress",updateProgressSuccess,updateProgressError,data);
+    }
+  }
+
+  const updateProgressSuccess = (res) => {
+    messageBoxHandle(true,"진척도 변경 완료",2000,'success');
+    updatePlanList();
+  }
+
+  const updateProgressError = () => {
+    _messageBoxHandle(true,"진척도 변경 중 오류가 발생했습니다.",2000,'error');
   }
 
   const deletePlanSuccess = () => {
     props.messageBoxHandle(true,"일정 삭제 완료",2000,'success');
     updatePlanList();
+  }
+
+  const deleteHandle = () => {
+    setConfirmDialogState(true);
+    setConfirmDialogInfo({
+      title : "삭제",
+      content : "위 일정을 정말 삭제하시겠습니까?"
+    });
+    setConfirmState(true);
+  }
+  
+  const updateProgress = () => {
+    setConfirmDialogState(true);
+    setConfirmDialogInfo({
+      title : "진척도 변경",
+      content : "정말 진척도를 변경하시겠습니까?"
+    });
+    setConfirmState(false);
+  }
+
+  const updateProgressState = (num) => {
+    progress.current.value = num;
   }
 
   useEffect(()=>{
@@ -142,17 +204,32 @@ export default function ShowSelectEvent(props) {
             <Typography className={classes.pos} color="textSecondary">
               {event ? event['start'] : null} ~ {event ? event['end'] : null}
             </Typography>
-            <PrettoSlider valueLabelDisplay="auto" style={{width:'65%'}} aria-label="pretto slider" defaultValue={20} />
-            <Button size="small" style={{float:'right'}} variant="outlined" color="secondary">변 경</Button>
+            <PrettoSlider valueLabelDisplay="auto" style={{width:'65%'}} aria-label="pretto slider" defaultValue={props['event'] ? props['event']['progress'] : null} onChange={(e,num)=>updateProgressState(num)}/>
+            <input type="hidden" ref={progress}/>
+            {props['event'] && props['event']['user']['id'] === localStorage.getItem('ID') ? 
+            (
+              <Button size="small" style={{float:'right'}} variant="outlined" color="secondary" onClick={updateProgress}>변 경</Button>
+            ) : 
+            null}
           </CardContent>
+          {props['event'] && props['event']['user']['id'] === localStorage.getItem('ID') ? 
+          (
           <CardActions style={{float:'right'}}>
             <Button size="small" onClick={()=>setUpdatePlanState(true)}>수정</Button>
-            <Button size="small" onClick={()=>setDeleteState(true)}>삭제</Button>
+            <Button size="small" onClick={deleteHandle}>삭제</Button>
           </CardActions>
+          ) : null}
         </Card>
       </Dialog>
       <UpdatePlan updatePlanList={updatePlanList} messageBoxHandle={messageBoxHandle} plan={props['event']} open={updatePlanState} handleClose={setUpdatePlanState}/>
-      <ConfirmDialog yseClick={deletePlan} title={"삭제"} content={"위 일정을 정말 삭제하시겠습니까?"}  open={deleteState} handleClose={()=>setDeleteState(false)}/>
+      <ConfirmDialog yseClick={yseClickHandle} title={confirmDialogInfo['title']} content={confirmDialogInfo['content']} open={confirmDialogState} handleClose={()=>setConfirmDialogState(false)}/>
+      <MessageBox
+          open={showMessageState}
+          content={MessageBoxState['content']}
+          level={MessageBoxState['level']}
+          time={MessageBoxState['time']}
+          handleClose={()=>setShowMessageState(false)}
+        />
     </div>
   );
 }
